Add files via upload
This is the version in Spring2023. I wished to make changes in Fall2023. However, I was too busy in my application. I hope I can use Graph Neural Network to better simulate the date and make synthetic prediction for my graduation debate. But I still confused whether this method works or not...